+++ test/SendGrid.php
@@ -12,14 +12,42 @@ class SendGridTest_SendGrid extends PHPUnit_Framework_TestCase
 
     public function testVersion()
     {
-        $this->assertEquals(SendGrid::VERSION, '3.0.0');
+        $this->assertEquals(SendGrid::VERSION, '3.2.0');
         $this->assertEquals(json_decode(file_get_contents('../composer.json'))->version, SendGrid::VERSION);
     }
 
-    public function testInitialization()
+    public function testInitWithUsernamePassword()
     {
         $sendgrid = new SendGrid('user', 'pass');
         $this->assertEquals('SendGrid', get_class($sendgrid));
+    }
+
+    public function testInitWithUsernamePasswordOptions()
+    {
+        $sendgrid = new SendGrid('user', 'pass', array('foo' => 'bar'));
+        $this->assertEquals('SendGrid', get_class($sendgrid));
+    }
+
+    public function testInitWithApiKey()
+    {
+        $sendgrid = new SendGrid('token1234');
+        $this->assertEquals('SendGrid', get_class($sendgrid));
+        $this->assertNull($sendgrid->apiUser);
+        $this->assertEquals($sendgrid->apiKey, 'token1234');
+    }
+
+    public function testInitWithApiKeyOptions()
+    {
+        $sendgrid = new SendGrid('token1234', array('foo' => 'bar'));
+        $this->assertEquals('SendGrid', get_class($sendgrid));
+    }
+
+    public function testInitWithProxyOption()
+    {
+        $sendgrid = new SendGrid('user', 'pass', array('proxy' => 'myproxy.net:3128'));
+        $this->assertEquals('SendGrid', get_class($sendgrid));
+        $options = $sendgrid->getOptions();
+        $this->assertTrue(isset($options['proxy']));
     }
 
     public function testDefaultURL()
@@ -57,6 +85,22 @@ class SendGridTest_SendGrid extends PHPUnit_Framework_TestCase
         $mockResponse = (object)array('code' => 400, 'raw_body' => "{'message': 'error', 'errors': ['Bad username / password']}");
 
         $sendgrid = m::mock('SendGrid[postRequest]', array('foo', 'bar'));
+        $sendgrid->shouldReceive('postRequest')->once()->andReturn($mockResponse);
+
+        $email = new SendGrid\Email();
+        $email->setFrom('bar@foo.com')
+            ->setSubject('foobar subject')
+            ->setText('foobar text')
+            ->addTo('foo@bar.com');
+
+        $response = $sendgrid->send($email);
+    }
+
+    public function testDisableSendGridException()
+    {
+        $mockResponse = (object)array('code' => 400, 'raw_body' => "{'message': 'error', 'errors': ['Bad username / password']}");
+
+        $sendgrid = m::mock('SendGrid[postRequest]', array('foo', 'bar', array('raise_exceptions' => false)));
         $sendgrid->shouldReceive('postRequest')->once()->andReturn($mockResponse);
 
         $email = new SendGrid\Email();